requirements.txt
black==24.4.2
httptools==0.6.1
langchain-chroma==0.1.2
langchain-community==0.2.7
langchain-upstage==0.1.7
pip==24.0
PyMuPDF==1.24.7
pysqlite3-binary==0.5.3
python-dotenv==1.0.1
streamlit==1.36.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
